refactor: move runtime caching to scripts
